+++ src/index.ts
@@ -1,7 +1,9 @@
 export * from './langfuse.js';
 export * from './langfusePublicApi.js';
+export * from './langfuseTrigger.js';
 export * from './nodeLogic.js';
 export * from './n8n-lite.js';
 
 export { Langfuse } from '../nodes/Langfuse/Langfuse.node.js';
+export { LangfuseTrigger } from '../nodes/LangfuseTrigger/LangfuseTrigger.node.js';
 export { LangfuseApi } from '../credentials/LangfuseApi.credentials.js';